Add a GOST 7.1–2003 style
Based on the Russian GOST R 7.0.5–2008 (numeric, sorted alphabetically, Russian) style by Dmitriy Rybakov, edited so that to implement GOST 7.1–2003. The latter is important for being a required use in formatting a bibliography for a doctoral thesis in Russia. The style is a clear alpha. Tested Article, Book, and InProceedings BibTeX entries only. Subject to further development.