Enclose year in citation in parentheses
When it is cited within the text, ie the citation itself not enclosed,
then the year needs to be parenthesized, so I made appropriate
corrections.

However, in this way, years will get enclosed when within
the parenthesized citation, too, which is inappropriate.

To correct for the latter, I would do the following regex:

`/ \((\d{4})\)([);])/`

Replacement:

`, $1$2`

It is assumed the citations within parentheses are always
separated from each other with a semicolon, not with a comma.

P. S. MS Word as far as I get it still doesn’t know regex.
So good thing that OO does.